New Version Updates-Model (minor) modifications

+++ Source.VS2022.C#.Template/iAppy.CCBS.Shared/Models/CbsCustomer/GetCustomerRelationshipsRequest.cs
@@ -8,7 +8,7 @@ namespace iApplyShared.Models
         #region Public Properties
 
         [DataMember]
-        public short? CustomerCBSCd { get; set; }
+        public int? CustomerCBSCd { get; set; }
 
         [DataMember]
         public int? CustomerCBSId { get; set; }
